backend/requirements.txt
# Backend API dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
sqlmodel==0.0.27
asyncpg==0.30.0
httpx==0.27.2
websockets==12.0

# YOLO/ML dependencies
ultralytics
opencv-python
numpy
torch
torchvision
Pillow
boto3
botocore
